image_generator/requirements.txt
requests>=2.25.1
pymongo>=3.12.0
python-dotenv>=0.19.0
Pillow>=8.3.1
tqdm>=4.62.2
replicate>=0.8.0
tenacity>=8.0.1 